refactor:  command paths 수정
