+++ Dockerfile.local
@@ -21,7 +21,7 @@ COPY src/ ./src/
 RUN npm run build
 
 # 로컬 테스트용: S3Uploader를 로컬 버전으로 교체
-RUN cp /build/dist/uploader/S3Uploader.local.js /build/dist/uploader/S3Uploader.js
+RUN cp /build/dist/aws/s3/S3Uploader.local.js /build/dist/aws/s3/S3Uploader.js
 
 # === 런타임 스테이지 ===
 FROM --platform=linux/amd64 public.ecr.aws/lambda/nodejs:22
